+++ src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './BigToolbar';
 export * from './CalendarPanel';
+export * from './DateEditor';
 export * from './DefaultViewEditor';
 export * from './MultiFieldEditor';
 export * from './TimeEditor';