+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated tutorial video (#190)
 - Added Formatting for event description (#182)
 - Updated Details drawer layout (#192)
+- Added support for additional Time Range sources (#179)
 
 ## 3.1.0 (2024-05-19)