+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added Formatting for event description (#182)
 - Updated Details drawer layout (#192)
 - Added support for additional Time Range sources (#179)
+- Update format and view (#188)
+- Added display dashboard annotations (#194)
 
 ## 3.1.0 (2024-05-19)