+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added Formatting for event description (#182)
 - Updated Details drawer layout (#192)
 - Added support for additional Time Range sources (#179)
-- Update format and view (#188)
+- Update event format (#188)
 - Updated Time Range in Agenda view (#196)
 - Added display dashboard annotations (#194)